+++ utils/tools.py
@@ -1,11 +1,9 @@
 # TOOLS.PY
 # A collection of miscellaneous functions to simplify common tasks in IntFicPy
 
-from . import thing
-from . import actor
-from . import score
-from . import travel
-from . import room
+from intficpy.things import thing, actor
+from intficpy.score import score
+from intficpy.travel import travel, room
 
 def isSerializableClassInstance(obj):
 	return isinstance(obj, thing.Thing) or isinstance(obj, actor.Actor) or isinstance(obj, score.Achievement) or isinstance(obj, score.Ending) or isinstance(obj, thing.Abstract) or isinstance(obj, actor.Topic) or isinstance(obj, travel.TravelConnector) or isinstance(obj, room.Room) or isinstance(obj, actor.SpecialTopic)  or isinstance(obj, actor.SaleItem) or isinstance(obj, score.HintNode) or isinstance(obj, room.RoomGroup) or isinstance(obj, score.Hint)
