+++ vocab/vocab.py
